Rename lang.c to call.c
