+++ src/rlang/rlang.h
@@ -78,6 +78,7 @@ extern sexp* r_shared_false;
 #include "attrib.h"
 #include "debug.h"
 #include "c-utils.h"
+#include "call.h"
 #include "cnd.h"
 #include "dict.h"
 #include "env.h"
@@ -86,7 +87,6 @@ extern sexp* r_shared_false;
 #include "export.h"
 #include "fn.h"
 #include "formula.h"
-#include "lang.h"
 #include "node.h"
 #include "parse.h"
 #include "quo.h"
